refactor: rename log dialog
