+++ assets/ui/log_dialog.ui
