+++ src/hd_active/ui/forms/ui_log_dialog.py
@@ -1,54 +1,22 @@
 # -*- coding: utf-8 -*-
 
 ################################################################################
-## Form generated from reading UI file 'log.ui'
+## Form generated from reading UI file 'log_dialog.ui'
 ##
-## Created by: Qt User Interface Compiler version 6.2.4
+## Created by: Qt User Interface Compiler version 6.7.2
 ##
 ## WARNING! All changes made in this file will be lost when recompiling UI file!
 ################################################################################
 
-from PySide6.QtCore import (
-    QCoreApplication,
-    QDate,
-    QDateTime,
-    QLocale,
-    QMetaObject,
-    QObject,
-    QPoint,
-    QRect,
-    QSize,
-    Qt,
-    QTime,
-    QUrl,
-)
-from PySide6.QtGui import (
-    QBrush,
-    QColor,
-    QConicalGradient,
-    QCursor,
-    QFont,
-    QFontDatabase,
-    QGradient,
-    QIcon,
-    QImage,
-    QKeySequence,
-    QLinearGradient,
-    QPainter,
-    QPalette,
-    QPixmap,
-    QRadialGradient,
-    QTransform,
-)
-from PySide6.QtWidgets import (
-    QApplication,
-    QDialog,
-    QPushButton,
-    QSizePolicy,
-    QTextBrowser,
-    QWidget,
-)
-
+from PySide6.QtCore import (QCoreApplication, QDate, QDateTime, QLocale,
+    QMetaObject, QObject, QPoint, QRect,
+    QSize, QTime, QUrl, Qt)
+from PySide6.QtGui import (QBrush, QColor, QConicalGradient, QCursor,
+    QFont, QFontDatabase, QGradient, QIcon,
+    QImage, QKeySequence, QLinearGradient, QPainter,
+    QPalette, QPixmap, QRadialGradient, QTransform)
+from PySide6.QtWidgets import (QApplication, QDialog, QPushButton, QSizePolicy,
+    QTextBrowser, QWidget)
 
 class Ui_LogDialog(object):
     def setupUi(self, LogDialog):
@@ -65,11 +33,10 @@ class Ui_LogDialog(object):
         self.retranslateUi(LogDialog)
 
         QMetaObject.connectSlotsByName(LogDialog)
-
     # setupUi
 
     def retranslateUi(self, LogDialog):
         LogDialog.setWindowTitle(QCoreApplication.translate("LogDialog", u"HD Active Log", None))
         self.ok_button.setText(QCoreApplication.translate("LogDialog", u"Ok", None))
-
     # retranslateUi
+
