+++ src/hd_active/ui/log_dialog.py
@@ -2,7 +2,7 @@ from PySide6.QtGui import QShowEvent
 from PySide6.QtWidgets import QDialog
 
 from src.hd_active.hd_active import HdActive
-from src.hd_active.ui.forms.log_ui import Ui_LogDialog
+from src.hd_active.ui.forms.ui_log_dialog import Ui_LogDialog
 
 
 class LogDialog(QDialog):
